refactor(client): reorganize `/session/ingame` components
