+++ apps/next-app/src/app/game/(in-game)/offline/page.tsx
@@ -1,7 +1,7 @@
 import dynamic from "next/dynamic"
 
 // components
-import { SessionLoader } from "@/components/session/ingame"
+import { SessionLoader } from "@/components/gameplay"
 
 export default dynamic(() => import("./offline-session-loader"), {
   ssr: false,
