+++ apps/next-app/src/components/gameplay/index.ts
@@ -1,0 +1,4 @@
+export { default as SessionHeader } from "./header/session-header"
+export { default as SessionFooter } from "./footer/session-footer"
+export { default as MemoryTable } from "./memory-table"
+export { default as SessionLoader } from "./session-loader"
