+++ apps/next-app/src/components/gameplay/memory-table.tsx
@@ -1,5 +1,5 @@
 // types
-import type { ClientSession, ClientSessionCard } from "@repo/schema/session"
+import type { ClientSessionCard } from "@repo/schema/session"
 
 // utils
 import { cn } from "@/lib/util"
@@ -7,18 +7,22 @@ import { cn } from "@/lib/util"
 // components
 import MemoryCard from "./memory-card"
 
+// hooks
+import { useSessionStore } from "@/components/provider/session-store-provider"
+
 type MemoryTableProps = {
-  session: ClientSession
   handleCardFlip: (clickedCard: ClientSessionCard) => void
 }
 
-const MemoryTable = ({ session, handleCardFlip }: MemoryTableProps) => {
+const MemoryTable = ({ handleCardFlip }: MemoryTableProps) => {
+  const session = useSessionStore((state) => state.session)
+
   return (
     <main className="flex-1 w-full p-4 flex justify-center items-center">
       <div className={cn("grid gap-4 w-full grid-cols-3", {
-        "max-w-3xl md:grid-cols-4": session.tableSize === 'SMALL',
-        "max-w-5xl sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6": session.tableSize === 'MEDIUM',
-        "max-w-7xl sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6 xl:grid-cols-8": session.tableSize === 'LARGE'
+        "max-w-3xl md:grid-cols-4": session.tableSize === "SMALL",
+        "max-w-5xl sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6": session.tableSize === "MEDIUM",
+        "max-w-7xl sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6 xl:grid-cols-8": session.tableSize === "LARGE"
       }
       )}>
         {session.cards.map((card) => (
