--- apps/next-app/src/components/session/ingame/index.ts
@@ -1,5 +1,0 @@
-export { default as MemoryTable } from "./memory-table"
-export { default as MemoryCard } from "./memory-card"
-export { default as SessionLoader } from "./session-loader"
-export { default as SessionFooter } from "./footer/session-footer"
-export { default as SessionHeader } from "./header/session-header"
